+++ apps/prairielearn/src/tests/helperDb.ts
@@ -1,17 +1,18 @@
 // @ts-check
-const pg = require('pg');
-const path = require('path');
-const util = require('util');
+import * as pg from 'pg';
+import * as path from 'path';
+import { promisify } from 'util';
 
-const sqldb = require('@prairielearn/postgres');
-const {
-  init: initMigrations,
+import * as sqldb from '@prairielearn/postgres';
+import {
+  init as initMigrations,
   initBatchedMigrations,
   SCHEMA_MIGRATIONS_PATH,
   stopBatchedMigrations,
-} = require('@prairielearn/migrations');
-const sprocs = require('../sprocs');
-const namedLocks = require('@prairielearn/named-locks');
+} from '@prairielearn/migrations';
+import * as sprocs from '../sprocs';
+import * as namedLocks from '@prairielearn/named-locks';
+import { Context } from 'mocha';
 
 const POSTGRES_USER = 'postgres';
 const POSTGRES_HOST = 'localhost';
@@ -36,12 +37,7 @@ const postgresTestUtils = sqldb.makePostgresTestUtils({
   },
 });
 
-/**
- *
- * @param {string} dbName
- * @param {boolean} runMigrations
- */
-async function runMigrationsAndSprocs(dbName, runMigrations) {
+async function runMigrationsAndSprocs(dbName: string, runMigrations: boolean): Promise<void> {
   const pgConfig = {
     user: POSTGRES_USER,
     database: dbName,
@@ -73,20 +69,18 @@ async function runMigrationsAndSprocs(dbName, runMigrations) {
   }
 
   await sqldb.setRandomSearchSchemaAsync('test');
-  await util.promisify(sprocs.init)();
+  await promisify(sprocs.init)();
 
   await stopBatchedMigrations();
   await namedLocks.close();
   await sqldb.closeAsync();
 }
 
-/**
- *
- * @param {string} dbName
- * @param {string} dbTemplateName
- * @param {boolean} dropFirst
- */
-async function createFromTemplate(dbName, dbTemplateName, dropFirst) {
+async function createFromTemplate(
+  dbName: string,
+  dbTemplateName: string,
+  dropFirst: boolean,
+): Promise<void> {
   await postgresTestUtils.createDatabase({
     dropExistingDatabase: dropFirst,
     database: dbName,
@@ -96,12 +90,12 @@ async function createFromTemplate(dbName, dbTemplateName, dropFirst) {
   });
 }
 
-async function closeSql() {
+async function closeSql(): Promise<void> {
   await namedLocks.close();
   await sqldb.closeAsync();
 }
 
-async function databaseExists(dbName) {
+async function databaseExists(dbName: string): Promise<boolean> {
   const client = new pg.Client(POSTGRES_INIT_CONNECTION_STRING);
   await client.connect();
   const result = await client.query(
@@ -112,15 +106,13 @@ async function databaseExists(dbName) {
   return existsResult;
 }
 
-async function setupDatabases() {
+async function setupDatabases(): Promise<void> {
   const templateExists = await databaseExists(POSTGRES_DATABASE_TEMPLATE);
-  const dbName = module.exports.getDatabaseNameForCurrentWorker();
-  if (templateExists) {
-    await createFromTemplate(dbName, POSTGRES_DATABASE_TEMPLATE, true);
-  } else {
-    await module.exports.createTemplate();
-    await createFromTemplate(dbName, POSTGRES_DATABASE_TEMPLATE, true);
+  const dbName = getDatabaseNameForCurrentWorker();
+  if (!templateExists) {
+    await createTemplate();
   }
+  await createFromTemplate(dbName, POSTGRES_DATABASE_TEMPLATE, true);
 
   // Ideally this would happen only over in `helperServer`, but we need to use
   // the same database details, so this is a convenient place to do it.
@@ -129,68 +121,60 @@ async function setupDatabases() {
   });
 }
 
-/**
- * @this {import('mocha').Context}
- */
-module.exports.before = async function before() {
+export async function before(this: Context): Promise<void> {
   // long timeout because DROP DATABASE might take a long time to error
   // if other processes have an open connection to that database
   this.timeout?.(20000);
   await setupDatabases();
-};
+}
 
 /**
  * This version will only (re)create the database with migrations; it will
  * then close the connection in sqldb. This is necessary for database
  * schema verification, where databaseDiff will set up a connection to the
  * desired database.
- *
- * @this {import('mocha').Context}
  */
-module.exports.beforeOnlyCreate = async function beforeOnlyCreate() {
+export async function beforeOnlyCreate(this: Context): Promise<void> {
   // long timeout because DROP DATABASE might take a long time to error
   // if other processes have an open connection to that database
   this.timeout?.(20000);
   await setupDatabases();
   await closeSql();
-};
+}
 
-/**
- * @this {import('mocha').Context}
- */
-module.exports.after = async function after() {
+export async function after(this: Context): Promise<void> {
   // long timeout because DROP DATABASE might take a long time to error
   // if other processes have an open connection to that database
   this.timeout?.(20000);
   await closeSql();
   await postgresTestUtils.dropDatabase();
-};
+}
 
-module.exports.createTemplate = async function createTemplate() {
+export async function createTemplate(): Promise<void> {
   await postgresTestUtils.createDatabase({
     dropExistingDatabase: true,
     database: POSTGRES_DATABASE_TEMPLATE,
     configurePool: false,
     prepare: () => runMigrationsAndSprocs(POSTGRES_DATABASE_TEMPLATE, true),
   });
-};
+}
 
-module.exports.dropTemplate = async function dropTemplate() {
+export async function dropTemplate(): Promise<void> {
   await closeSql();
   await postgresTestUtils.dropDatabase({
     database: POSTGRES_DATABASE_TEMPLATE,
     // Always drop the template regardless of PL_KEEP_TEST_DB env
     force: true,
   });
-};
+}
 
-module.exports.resetDatabase = async function resetDatabase() {
+export async function resetDatabase(): Promise<void> {
   await postgresTestUtils.resetDatabase();
-};
+}
 
-module.exports.getDatabaseNameForCurrentWorker = function getDatabaseNameForCurrentWorker() {
+export function getDatabaseNameForCurrentWorker(): string {
   return postgresTestUtils.getDatabaseNameForCurrentMochaWorker();
-};
+}
 
 class RollbackTransactionError extends Error {
   constructor() {
@@ -207,7 +191,7 @@ class RollbackTransactionError extends Error {
  * The current transaction will not be propagated across network calls, so
  * use this carefully.
  */
-module.exports.runInTransactionAndRollback = async function runInTransactionAndRollback(fn) {
+export async function runInTransactionAndRollback(fn: () => Promise<void>): Promise<void> {
   await sqldb
     .runInTransactionAsync(async () => {
       await fn();
@@ -219,4 +203,4 @@ module.exports.runInTransactionAndRollback = async function runInTransactionAndR
       }
       throw err;
     });
-};
+}